rag-system/requirements.txt
autogen>=0.7.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
openai>=1.12.0
python-multipart>=0.0.9 